Move pkg/version to internal/version
This change is part of our ongoing effort to move public packages that are not intended for external use into the /internal directory. This prevents accidental external dependencies on our implementation details.

- Moved all files from /pkg/version to /internal/version
- Updated all import paths throughout the codebase
- No functional changes
