+++ cmd/tracegen/main.go
@@ -27,7 +27,7 @@ import (
 	"github.com/jaegertracing/jaeger/internal/jaegerclientenv2otel"
 	"github.com/jaegertracing/jaeger/internal/tracegen"
 	"github.com/jaegertracing/jaeger/pkg/otelsemconv"
-	"github.com/jaegertracing/jaeger/pkg/version"
+	"github.com/jaegertracing/jaeger/internal/version"
 )
 
 var flagAdaptiveSamplingEndpoint string
