+++ internal/version/command_test.go
